ifplugd: Allow ENODEV when upping interface for -M
Allow ENODEV error in up_iface. This handles an edge case running ifplugd
with the option to detect device hot plugging if the device disappears after
the check of G.iface_exists.

If the interface goes away, set G.iface_exists and return. If -M is not set,
then other places in the code will detect the missing interface and exit (the
original behavior).

Signed-off-by: Jate Sujjavanich <jatedev@gmail.com>